feat: move cell ui to teka-tekit
