+++ packages/naive_cross_word/pubspec.yaml
@@ -10,6 +10,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  teka_tekit: ^0.1.1
 
 dev_dependencies:
   flutter_test:
