--- packages/teka_tekit/lib/src/teka_tekit.dart
@@ -1,7 +1,0 @@
-/// {@template teka_tekit}
-/// teka_teki_teku UI KIT
-/// {@endtemplate}
-class TekaTekit {
-  /// {@macro teka_tekit}
-  const TekaTekit();
-}
