+++ packages/teka_tekit/lib/teka_tekit.dart
@@ -1,4 +1,4 @@
 /// teka_teki_teku UI KIT
 library;
 
-export 'src/teka_tekit.dart';
+export 'src/components/cells/cells.dart';
