+++ packages/teka_tekit/test/src/teka_tekit_test.dart
@@ -1,12 +1,11 @@
 // ignore_for_file: prefer_const_constructors
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:teka_tekit/teka_tekit.dart';
 
 void main() {
   group('TekaTekit', () {
-    test('can be instantiated', () {
-      expect(TekaTekit(), isNotNull);
-    });
+    // test('can be instantiated', () {
+    //   expect(TekaTekit(), isNotNull);
+    // });
   });
 }
